requirements.txt
fastapi
deta
uvicorn
jinja2
aiofiles
paho-mqtt
tinydb
